requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0
aiogram>=3.5.0
PyQt5>=5.15.0
httpx>=0.27.0
aiohttp>=3.9.0
python-dateutil>=2.9.0
supabase>=2.0.0
